We suck at initializing things
iTem was never initialized, so it was going bonkers
we also finished the extinguish logic

it works, assuming you're 1m away from candle

+++ src/shared/PID.hpp
@@ -13,7 +13,7 @@ class PID{
            iTerm;
 
     float setPoint;
-    float output;
+    int output;
 
     bool feedForward = false;
 };
